+++ lib/presentation/group_post/view/view.dart
@@ -1,1 +1,3 @@
+export 'group_member_list_view_widget.dart';
 export 'group_post_view.dart';
+export 'group_post_view_widget.dart';
